fix(egress): cache nameserver exempt addrs and split exempt tests
